backend/app/core/auth.py

from fastapi import Depends, HTTPException, status
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from jose import jwt, JWTError
from app.db import supabase
from typing import Optional, Dict

# Security scheme for JWT token
security = HTTPBearer()

class AuthError(Exception):
    def __init__(self, message: str):
        self.message = message
        super().__init__(self.message)

async def get_current_user(
    credentials: HTTPAuthorizationCredentials = Depends(security)
) -> Dict:
    """
    Validate the JWT token and return the user data.
    """
    try:
        # This is a simple placeholder for actual JWT validation
        # In a real implementation, you would verify the token with Supabase
        token = credentials.credentials
        
        # For now, just check if the token exists and has a valid format
        if not token:
            raise AuthError("Missing token")
        
        # In a real implementation, verify the token and extract user data
        # Example: user_data = supabase.auth.get_user(token)
        
        # For now, returning a placeholder user
        # In reality, you would decode and verify the JWT
        user_id = "placeholder_user_id"
        
        return {"user_id": user_id}
    
    except JWTError as e:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail=f"Invalid authentication credentials: {str(e)}",
            headers={"WWW-Authenticate": "Bearer"},
        )
    except AuthError as e:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail=e.message,
            headers={"WWW-Authenticate": "Bearer"},
        )
