backend/requirements.txt

fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.5.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
supabase==2.0.3
httpx==0.25.1
python-jose==3.3.0
python-multipart==0.0.6
